重构run文件，现在可以在命令行--config_pth后或py中的配置argparse直接输入ini文件位置即可
