--- config/EpinionsLightGCN.ini
@@ -1,39 +1,0 @@
-[TRAIN]
-epoch = 150
-device = cuda
-log_pth = ./log/
-save_pth = ./save/
-random_seed = 78912
-mode = debug
-eval_step = 5
-warm_epoch = 150
-task = Rate
-
-[OPTIM]
-optimizer = AdamW
-lr_scheduler = CosineAnnealingWarmRestarts
-lr_adapt_step = 10
-learning_rate = 1e-1
-weight_decay = 1e-4
-
-[LOSS]
-loss_name = BPRLoss
-
-[MODEL]
-model_name = LightGCN
-# 需要预测的user数量，link与rate任务需要预测的user应该是相同的
-pred_user_num = 19044
-# link任务中出现但是无需预测的user
-total_user_num = 45180
-item_num = 53746
-layer_num = 2
-embedding_size = 10
-
-[DATA]
-data_name = Epinions
-train_neg_num = 4
-neg_num = 100
-
-[METRIC]
-ks = [3, 5, 10]
-metric_name = ['HR', 'nDCG']
