+++ config/LightGCN.ini
@@ -6,7 +6,7 @@ save_pth = ./save/
 random_seed = 78912
 mode = debug
 eval_step = 5
-warm_epoch = 25
+warm_epoch = 150
 task = Rate
 
 [OPTIM]
@@ -14,7 +14,7 @@ optimizer = AdamW
 lr_scheduler = CosineAnnealingWarmRestarts
 lr_adapt_step = 10
 learning_rate = 1e-1
-weight_decay = 1e-4
+weight_decay = 1e-3
 
 [LOSS]
 loss_name = BPRLoss
@@ -22,15 +22,16 @@ loss_name = BPRLoss
 [MODEL]
 model_name = LightGCN
 # 需要预测的user数量，link与rate任务需要预测的user应该是相同的
-pred_user_num = 1598
+pred_user_num = 19044
 # link任务中出现但是无需预测的user
-total_user_num = 19790
-item_num = 24985
+total_user_num = 45180
+item_num = 53746
 layer_num = 2
 embedding_size = 10
 
 [DATA]
 data_name = Epinions
+train_neg_num = 4
 neg_num = 100
 
 [METRIC]
