+++ run.py
@@ -1,13 +1,14 @@
 #!/usr/bin/env python
 # -*- encoding: utf-8 -*-
 """
-@File    :   run_TrustSVD.py
+@File    :   run.py
 @Contact :   xansar@ruc.edu.cn
 
 @Modify Time      @Author    @Version    @Desciption
 ------------      -------    --------    -----------
-2022/10/25 13:51   zxx      1.0         None
+2022/10/26 16:22   zxx      1.0         None
 """
+
 import os
 
 import torch
@@ -23,11 +24,16 @@ from trainer import *
 
 from configparser import ConfigParser
 
+CommonModel = ['MF']
+GCNModel = ['LightGCN', 'TrustSVD']
+
 class MyConfigParser(ConfigParser):
-    def __init__(self,defaults=None):
+    def __init__(self, defaults=None):
         super(MyConfigParser, self).__init__()
+
     def optionxform(self, optionstr):
         return optionstr
+
 
 def get_config(config_pth):
     config = MyConfigParser()
@@ -46,7 +52,7 @@ def parse_args():
     # Parses the arguments.
     parser = argparse.ArgumentParser(description="Run Model.")
     parser.add_argument('--config_pth', type=str, default='TrustSVD.ini',
-                    help='Choose config')
+                        help='Choose config')
     return parser.parse_args()
 
 
@@ -57,28 +63,20 @@ def run(config_pth):
     data_name = config['DATA']['data_name']
     model_name = config['MODEL']['model_name']
     task = config['TRAIN']['task']
-    dataset_name = data_name + task + model_name
-    num_workers = eval(config['DATA']['num_workers'])
-    # data_pth = '../data/ExtendedEpinions/splited_data/MFModel/val'
-    dataset_dict = {}
-    dataloader_dict = {}
-    for m in ['train', 'val', 'test']:
-        data_pth = os.path.join('./data', data_name, 'splited_data', f'{m}')
-        dataset_dict[m] = eval(dataset_name)(data_pth=data_pth, config=config)
 
-        batch_size = eval(config['DATA'][f'{m}_batch_size'])
-        is_shuffle = True if m == 'train' else False
-        dataloader_dict[m] = DataLoader(
-            dataset_dict[m],
-            batch_size=batch_size,
-            shuffle=is_shuffle,
-            num_workers=num_workers,
-            pin_memory=True,
-            collate_fn=TrustSVD_collate_fn
-        )
-    model = eval(model_name + 'Model')(config)
+    dataset_name = task + model_name
 
-    ## optimizer
+    dataset = eval(dataset_name)(config)
+    if model_name in CommonModel:
+        model = eval(model_name + 'Model')(config)
+    elif model_name in GCNModel:
+        etype = dataset[0].etypes
+        model = eval(model_name + 'Model')(config, etype)
+    else:
+        raise KeyError("Wrong Model Name!!!")
+    # model.apply(weight_init)
+
+    # optimizer
     lr = eval(config['OPTIM']['learning_rate'])
     weight_decay = eval(config['OPTIM']['weight_decay'])
     optimizer_name = 'torch.optim.' + config['OPTIM']['optimizer']
@@ -86,34 +84,25 @@ def run(config_pth):
     lr_scheduler_name = 'torch.optim.lr_scheduler.' + config['OPTIM']['lr_scheduler']
     lr_adapt_step = eval(config['OPTIM']['lr_adapt_step'])
     lr_scheduler = eval(lr_scheduler_name)(optimizer, lr_adapt_step)
-
-    ## loss func
+    # loss func
     loss_name = config['LOSS']['loss_name']
-    lamda = eval(config['OPTIM']['lamda'])
-    lamda_t = eval(config['OPTIM']['lamda_t'])
-    loss_func = eval(loss_name)(reduction='mean', lamda=lamda, lamda_t=lamda_t)
-
-    ## metric
+    loss_func = eval(loss_name)(reduction='mean')
+    # metric
     ks = eval(config['METRIC']['ks'])
     metric_name = eval(config['METRIC']['metric_name'])
     metric = eval(model_name + 'Metric')(ks=ks, task=task, metric_name=metric_name)
-
-    ## trainer
+    # trainer
     trainer = eval(model_name + 'Trainer')(
         model=model,
         loss_func=loss_func,
         optimizer=optimizer,
         lr_scheduler=lr_scheduler,
         metric=metric,
-        dataloader_dict=dataloader_dict,
+        dataset=dataset,
         config=config
     )
-    # from line_profiler import LineProfiler
-    # p = LineProfiler()
-    # p_wrap = p(trainer.train)
-    # p_wrap()
-    # p.print_stats()
     trainer.train()
+
 
 if __name__ == '__main__':
     args = parse_args()
